# append, append_row, append_column: methods for appending an array or single rows and columns