--- a/src/impl_methods.rs
+++ b/src/impl_methods.rs
@@ -873,7 +873,7 @@ where
     /// ```
     pub fn select(&self, axis: Axis, indices: &[Ix]) -> Array<A, D>
     where
-        A: Copy,
+        A: Clone,
         S: Data,
         D: RemoveAxis,
     {