For owned arrays specifically (Array), allow appending rows and/or columns or 
whole arrays (along a given axis).

**New methods:**

- `.append_row(ArrayView1<A>) -> Result`
- `.append_column(ArrayView1<A>) -> Result`
- `.append(axis, ArrayView<A, D>) -> Result`
- `.move_into(impl Into<ArrayViewMut>)`

**New features:**

- `stack`, `concatenate` and `.select(...)` now support Clone elements
  (previously only Copy).

The axis we append along should be a "growing axis", i.e the axis with greatest,
and positive, stride. However if that axis is of length 0 or 1, we can always
convert it to the new growing axis.

These methods automatically move the whole array to a compatible memory
layout for appending, if needed.

The examples show that empty arrays are valid both ways (both memory
layouts) - and that might be the easiest way to use these methods for the users.

There were quite a few corner cases in the implementation in this PR, but hopefully
they should all be dealt with in a moderately clean way and with little duplication.

Fixes #269