@bluss 2021-03-05
Note to others: there already exists other ways to build arrays row-by-row or column-by-column: the stack and concat functions.