@jturner314 2021-03-06
Functionality like this would be really nice. I think a slightly different API would be more generally useful:

```rust
impl<A, D> Array<A, D>
where
    D: Dimension,
{
    /// Appends the subview to the specified axis.
    ///
    /// The memory layout of the array will be changed if necessary (by copying the data)
    /// so that subsequent calls to `.append_to_axis()` with the same axis will be fast.
    /// This occurs in at least the following cases:
    ///
    /// - If the outermost (i.e. largest absolute stride) axis of the array is not `axis`.
    /// - If the array is discontiguous.
    /// - If the stride for `axis` is negative.
    ///
    /// So, calling `.append_to_axis()` with varying values of `axis` will likely be very slow, but
    /// after one call to `.append_to_axis()`, subsequent calls to `.append_to_axis()` with the
    /// *same* axis will be relatively fast.
    ///
    /// **Errors** if the axis lengths of `subview` do not match the correponding axis lengths of `self`.
    pub fn append_to_axis<S2>(&mut self, axis: Axis, subview: &ArrayBase<S2, D::Smaller>) -> Result<(), ShapeError>
    where
        A: Clone,
        S2: Data<Elem = A>,
    {}
}
```

In other words, it would be nice for the operation to work for n-dimensional arrays of arbitrary layout. The layout of the array would be changed as necessary so that the specified axis would be the outermost axis, so that repeated calls to `.append_to_axis()` with the same axis would be relatively fast. The remaining question with this design is how to handle the layout of the other axes. We could decide to (1) preserve the existing order, (2) use the order in the subview being appended, or (3) something in-between. I'd suggest option (2), because I think in most cases `.append_to_axis()` will be called repeatedly with the same axis and subviews of the same layout, and because it would simplify the handling of drops if the `.clone()` of one of the elements in `subview` panics (since the new elements would be added in order, like in this PR).