@bluss 2021-03-06
Yeah that would be more useful, but the complexity is higher. Do we have an easy way to get to the implementation?

How to do the trick where we make `axis` the largest stride axis? If this is enough then maybe it's not too complex to implement:

```rust
a.swap_axis(0, axis);
let a = a.as_standard_layout().into_owned();
a.swap_axis(0, axis);
// axis is now the largest stride axis; and with that axis ignored, the rest is c-contig
```