@jturner314 2021-03-07
I suppose I'd suggest something like the following:

```rust
impl<A, D> Array<A, D>
where
    D: Dimension,
{
    /// Rearranges the layout if necessary (by copying the data) such that:
    ///
    /// - if the specified axis has length > 1, it is the outermost axis and
    ///   has a positive stride
    /// - the other axes are in standard layout
    /// - the array is contiguous
    ///
    /// # Example
    ///
    /// ```rust
    /// use ndarray::{Array4, Axis};
    ///
    /// let mut arr = Array4::<u8>::zeros([2, 3, 4, 5]);
    /// assert_eq!(arr.shape(), &[2, 3, 4, 5]);
    /// assert_eq!(arr.strides(), &[60, 20, 5, 1]);
    ///
    /// arr.make_axis_outer_pos(Axis(2));
    /// assert_eq!(arr.shape(), &[2, 3, 4, 5]);
    /// assert_eq!(arr.strides(), &[15, 5, 30, 1]);
    /// ```
    pub(crate) fn make_axis_outer_pos(&mut self, axis: Axis)
    where
        A: Clone,
        D: RemoveAxis,
    {
        let ax = axis.index();
        if self.is_contiguous()
            && is_standard_layout(&self.dim.remove_axis(axis), &self.strides.remove_axis(axis))
            && (self.len_of(axis) <= 1
                || axis == self.dim.max_stride_axis(&self.strides) && self.stride_of(axis) > 0)
        {
            return;
        }
        self.dim.slice_mut()[..=ax].rotate_right(1);
        self.strides.slice_mut()[..=ax].rotate_right(1);
        *self = Array::from_shape_vec(self.dim.clone(), self.iter().cloned().collect()).unwrap();
        self.dim.slice_mut()[..=ax].rotate_left(1);
        self.strides.slice_mut()[..=ax].rotate_left(1);
    }
}
```

(The existing `is_standard_layout` function would need to be moved out of the `.is_standard_layout()` method and into e.g. `dimension/mod.rs`.)

In principle, we could move the elements instead of cloning them, although that would require some unsafe code.

This is basically the same approach as what you wrote above, but uses a rotate instead of a swap so that the other axes are in standard layout. This would make it easy to append the subview since the layout would match the order of `subview.iter()`. Note that if `axis` has length 1 or any of the axes has length 0, we'd need to be careful to make sure the strides are correct after the subview is appended (since the stride of a length 1 axis can be anything, and if there's a length 0 axis, then the strides of the other axes can be 0).