@jturner314 2021-04-01
I'd suggest something like the following:

1. Check that `subview` has the correct shape.
2. Call `self.make_axis_outer_pos(axis)` to adjust the layout of `self`.
3. (Debug) assert that `self.data.ptr == self.ptr`. (`make_axis_outer_pos` should ensure that this is the case.)
4. Call `let v = self.data.take_as_vec();` to convert the `OwnedRepr` to a `Vec`.
5. Call `v.extend(subview.iter().cloned())`.
6. Convert `v` back into an `OwnedRepr` and assign it to `self.data`.
7. Update `self.ptr` to point at the new data.
8. Update `self.shape` by adding 1 to the length of the axis.
9. If the original axis length was <= 1, then make sure the new stride of the axis is correct, since the stride of an axis with length <= 1 could be almost anything. (This step could be performed here, or it could be handled with a modification to `make_axis_outer_pos`.)

This way, the tricky stuff is handled by `Vec::extend`, which AFAIK handles panics without leaks. I'm curious how you were thinking about implementing it, and where the leak could occur.