@bluss 2021-04-01
Hm, my implementation uses Zip to make sure shape is traversed in the right order. To use Vec::extend we need to guarantee the array is C-contig except for the axis we are extending, which is a further requirement. So I started a bit different. I'd ideally see that we don't depend on Vec at all - looking forward to future custom allocation, but that concern is just a minor thing in the details here.