@bluss 2021-04-08
Benchmark:

before:

```text
test select_1d    ... bench:       1,339 ns/iter (+/- 20)
test select_axis0 ... bench:       1,204 ns/iter (+/- 21)
test select_axis1 ... bench:      11,136 ns/iter (+/- 35)
```

after

```text
test select_1d    ... bench:         110 ns/iter (+/- 2)
test select_axis0 ... bench:       1,169 ns/iter (+/- 20)
test select_axis1 ... bench:      21,087 ns/iter (+/- 176)
```

`.select()` used `concatenate` before, and select uses `.append()` after this PR.